+++ src/main/java/ru/maximkulikov/goodgame/api/chatmodels/ResJoin.java
@@ -31,10 +31,10 @@ public class ResJoin implements ResChatObject {
     private int smilePeka;
 
     @JsonProperty("clients_in_channel")
-    private Long clientsInChannel;
+    private Integer clientsInChannel;
 
     @JsonProperty("users_in_channel")
-    private Long usersInChannel;
+    private Integer usersInChannel;
 
     @JsonProperty("user_id")
     private String userId;
